+++ src/util/mod.rs
@@ -1,5 +1,6 @@
 //! Utility functions for the Shulkerbox project.
 
+mod command_collection;
 pub mod compile;
 mod extendable_queue;
 mod macro_string;
@@ -10,3 +11,6 @@ pub use extendable_queue::ExtendableQueue;
 
 #[doc(inline)]
 pub use macro_string::{MacroString, MacroStringPart};
+
+#[doc(inline)]
+pub use command_collection::CommandCollection;